chore(src): rename components - change the name of component folders to be in camelcase - rename root components to index.jsx - update paths for components with the new component names - update tests with the new component naming - update test snapshots [Delivers #159647999]
