+++ src/components/ImageLink/__tests__/ImageLink.test.js
@@ -1,17 +1,22 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import SearchBar from '../SearchBar';
+import ImageLink from '../index';
 
 // describe what we are testing
 describe("Render SearchBar component", () => {
+  const props = {
+    imageSrc:'',
+    altText:'',
+    imageClass:''
+  };
   // make our assertions and what we expect to happen
   it("should match snapshot", () => {
-    const wrapper = shallow(<SearchBar />);
+    const wrapper = shallow(<ImageLink {...props}/>);
     expect(wrapper).toMatchSnapshot();
   });
 
-  it("should render the searchbar as expected", () => {
-    const wrapper = shallow(<SearchBar />);
+  it("should render the ImageLink component as expected", () => {
+    const wrapper = shallow(<ImageLink {...props}/>);
     expect(wrapper.length).toBe(1);
   });
 });
