+++ src/components/NotificationPane/__tests__/NotificationPane.test.js
