+++ src/components/NotificationPane/_notificationContainer.scss
