+++ src/components/NotificationPane/_notificationHeader.scss
