+++ src/components/NotificationPane/_notificationItem.scss
