+++ src/components/NotificationPane/_notificationPane.scss
