+++ src/components/RequestPanelHeader/index.jsx
@@ -1,8 +1,9 @@
 import React, { PureComponent } from 'react';
-import Header from '../header/Header';
-import ButtonGroup from '../button-group/ButtonGroup';
+import Header from '../Header';
+import ButtonGroup from '../ButtonGroup';
 import HeaderPagination from '../Pagination/HeaderPagination';
 import './Request.scss';
+
 
 class RequestPanelHeader extends PureComponent {
   render() {
