+++ src/components/Requests/__tests__/Requests.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Requests from '../Requests';
+import Requests from '../index';
 
 const props = {
   requests: [
