+++ src/components/SearchBar/__tests__/SearchBar.test.js
@@ -1,22 +1,17 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import ImageLink from '../ImageLink';
+import SearchBar from '../index';
 
 // describe what we are testing
 describe("Render SearchBar component", () => {
-  const props = {
-    imageSrc:'',
-    altText:'',
-    imageClass:''
-  };
   // make our assertions and what we expect to happen
   it("should match snapshot", () => {
-    const wrapper = shallow(<ImageLink {...props}/>);
+    const wrapper = shallow(<SearchBar />);
     expect(wrapper).toMatchSnapshot();
   });
 
-  it("should render the ImageLink component as expected", () => {
-    const wrapper = shallow(<ImageLink {...props}/>);
+  it("should render the searchbar as expected", () => {
+    const wrapper = shallow(<SearchBar />);
     expect(wrapper.length).toBe(1);
   });
 });
