+++ src/views/App/__tests__/Request.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import RequestPanelHeader from '../../../components/RequestPanelHeader/RequestPanelHeader';
+import RequestPanelHeader from '../../../components/RequestPanelHeader';
 
 it('should render request header correctly', () => {
     const tree = renderer.create(<RequestPanelHeader />).toJSON();
