+++ src/views/RequestsPage/RequestsPage.jsx
@@ -1,12 +1,12 @@
 import React, { Component } from  'react';
-import Requests from '../../components/Requests/Requests';
+import Requests from '../../components/Requests';
 import Pagination from '../../components/Pagination/Pagination';
 import requestsData from '../../components/Requests/requestsData';
-import NavBar from '../../components/nav-bar/NavBar';
+import NavBar from '../../components/Navbar';
 import upic from '../../images/upic.svg';
 import './_index.scss';
-import NotificationPane from '../../components/notification-pane/NotificationPane';
-import RequestPanelHeader from '../../components/RequestPanelHeader/RequestPanelHeader';
+import NotificationPane from '../../components/NotificationPane/NotificationPane';
+import RequestPanelHeader from '../../components/RequestPanelHeader';
 
 class RequestsPage extends Component {
   // FIX: Remove console statement and replace with actual function
